fix: bridge header includes Hypervisor/Hypervisor.h directly, remove HypervisorBridge.c, fix HV_SUCCESS/HV_REG casts, fix onChange signature

+++ iCore/VM/HypervisorWrapper.swift
@@ -34,7 +34,7 @@ final class HypervisorWrapper {
     /// Returns false gracefully on simulator / devices without the entitlement.
     func createVM() -> Bool {
         let ret = hv_vm_create(nil)
-        guard ret == Int32(HV_SUCCESS) else {
+        guard ret == HV_SUCCESS else {
             console?.emit("[HV] hv_vm_create returned \(ret) — demo mode active.\n")
             return false
         }
@@ -58,7 +58,7 @@ final class HypervisorWrapper {
         let gpa = UInt64(0x4000_0000)
         let flags = UInt64(HV_MEMORY_READ) | UInt64(HV_MEMORY_WRITE) | UInt64(HV_MEMORY_EXEC)
         let mapRet = hv_vm_map(mem, gpa, alignedSz, flags)
-        guard mapRet == Int32(HV_SUCCESS) else {
+        guard mapRet == HV_SUCCESS else {
             console?.emit("[HV] hv_vm_map failed (\(mapRet)) — demo mode active.\n")
             return false
         }
@@ -70,7 +70,7 @@ final class HypervisorWrapper {
     func createVCPU() -> Bool {
         var exitPtr: UnsafeMutablePointer<hv_vcpu_exit_t>?
         let ret = hv_vcpu_create(&vcpu, &exitPtr, nil)
-        guard ret == Int32(HV_SUCCESS) else {
+        guard ret == HV_SUCCESS else {
             console?.emit("[HV] hv_vcpu_create failed (\(ret)).\n")
             return false
         }
@@ -93,8 +93,8 @@ final class HypervisorWrapper {
                .copyMemory(from: src.baseAddress!, byteCount: data.count)
         }
         let pc = UInt64(0x4000_0000) + UInt64(kernelOffset)
-        hv_vcpu_set_reg(vcpu, UInt32(HV_REG_PC), pc)
-        hv_vcpu_set_reg(vcpu, UInt32(HV_REG_CPSR), 0x3C5)    // EL1h, IRQ/FIQ masked
+        hv_vcpu_set_reg(vcpu, HV_REG_PC, pc)
+        hv_vcpu_set_reg(vcpu, HV_REG_CPSR, 0x3C5)    // EL1h, IRQ/FIQ masked
         console?.emit("[HV] Kernel loaded: \(data.count) bytes, PC=0x\(String(pc, radix: 16))\n")
         return true
     }
@@ -108,8 +108,8 @@ final class HypervisorWrapper {
             mem.advanced(by: 0).copyMemory(from: src.baseAddress!, byteCount: src.count)
         }
         let pc = UInt64(0x4000_0000)
-        hv_vcpu_set_reg(vcpu, UInt32(HV_REG_PC), pc)
-        hv_vcpu_set_reg(vcpu, UInt32(HV_REG_CPSR), 0x3C5)
+        hv_vcpu_set_reg(vcpu, HV_REG_PC, pc)
+        hv_vcpu_set_reg(vcpu, HV_REG_CPSR, 0x3C5)
         console?.emit("[HV] Test binary loaded (\(binary.count) bytes).\n")
     }
 
@@ -121,24 +121,24 @@ final class HypervisorWrapper {
             let uartBase = UInt64(0x0900_0000)
             while self.running {
                 let ret = hv_vcpu_run(self.vcpu)
-                guard ret == Int32(HV_SUCCESS) else {
+                guard ret == HV_SUCCESS else {
                     onExit("[HV] vcpu_run error: \(ret)\n")
                     break
                 }
                 // Without the full exit struct definition we advance PC
                 // and handle writes to our UART MMIO address heuristically.
                 var pc: UInt64 = 0
-                hv_vcpu_get_reg(self.vcpu, UInt32(HV_REG_PC), &pc)
+                hv_vcpu_get_reg(self.vcpu, HV_REG_PC, &pc)
                 // Detect UART write: if x0 is printable ASCII and pc is near UART
                 var x0: UInt64 = 0
-                hv_vcpu_get_reg(self.vcpu, UInt32(HV_REG_X0), &x0)
+                hv_vcpu_get_reg(self.vcpu, HV_REG_X0, &x0)
                 if x0 >= 0x20 && x0 < 0x7F {
                     self.console?.processByte(UInt8(x0 & 0xFF))
                 } else if x0 == 0x0A {
                     self.console?.processByte(0x0A)
                 }
                 // Advance PC past the faulting instruction
-                hv_vcpu_set_reg(self.vcpu, UInt32(HV_REG_PC), pc &+ 4)
+                hv_vcpu_set_reg(self.vcpu, HV_REG_PC, pc &+ 4)
 
                 // WFI / halt sentinel: x0 == 0 and pc hasn't changed
                 if x0 == 0 {
